Check if the key container exists before removing it

+++ USBHelperLauncher/Program.cs
@@ -221,8 +221,11 @@ namespace USBHelperLauncher
                     string rsaPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "Microsoft", "Crypto", "RSA", userSID);
                     string keyContainer = Directory.GetFiles(rsaPath).Where(n => Path.GetFileName(n).ToLower().StartsWith(hashString)).FirstOrDefault();
 
-                    File.Delete(keyContainer);
-                    logger.WriteLine("Removed broken key container.");
+                    if (keyContainer != default(string))
+                    {
+                        File.Delete(keyContainer);
+                        logger.WriteLine(string.Format("Removed broken key container (Name: \"{0}\", Hash: {1}).", keyContainerName, hashString));
+                    }
                 }
                 
             }
